trackerr_v1/requirements.txt
acme==3.0.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
configobj==5.0.9
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.9.0
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dkimpy==1.1.5
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
inflection==0.5.1
jaraco.classes==3.3.1
jeepney==0.8.0
josepy==1.14.0
keyring==24.3.1
keyrings.alt==5.0.0
kombu==5.3.7
lxml==5.1.0
more-itertools==10.2.0
packaging==24.0
parsedatetime==2.6
pillow==10.2.0
premailer==3.10.0
prompt-toolkit==3.0.46
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
SecretStorage==3.3.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yagmail==0.15.293
zipp==3.20.2
acme==3.0.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
configobj==5.0.9
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.9.0
daphne==4.1.2
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dkimpy==1.1.5
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==24.7.2
inflection==0.5.1
jaraco.classes==3.3.1
jeepney==0.8.0
josepy==1.14.0
keyring==24.3.1
keyrings.alt==5.0.0
kombu==5.3.7
lxml==5.1.0
more-itertools==10.2.0
packaging==24.0
parsedatetime==2.6
pillow==10.2.0
premailer==3.10.0
prompt-toolkit==3.0.46
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
SecretStorage==3.3.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yagmail==0.15.293
zipp==3.20.2
zope.interface==7.2
acme==3.0.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
configobj==5.0.9
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.9.0
daphne==4.1.2
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dkimpy==1.1.5
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==24.7.2
inflection==0.5.1
jaraco.classes==3.3.1
jeepney==0.8.0
josepy==1.14.0
keyring==24.3.1
keyrings.alt==5.0.0
kombu==5.3.7
lxml==5.1.0
more-itertools==10.2.0
msgpack==1.1.0
packaging==24.0
parsedatetime==2.6
pillow==10.2.0
premailer==3.10.0
prompt-toolkit==3.0.46
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
SecretStorage==3.3.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yagmail==0.15.293
zipp==3.20.2
zope.interface==7.2
acme==3.0.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.37.38
botocore==1.37.38
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
configobj==5.0.9
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.9.0
daphne==4.1.2
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dkimpy==1.1.5
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==24.7.2
inflection==0.5.1
jaraco.classes==3.3.1
jeepney==0.8.0
jmespath==1.0.1
josepy==1.14.0
keyring==24.3.1
keyrings.alt==5.0.0
kombu==5.3.7
lxml==5.1.0
more-itertools==10.2.0
msgpack==1.1.0
packaging==24.0
parsedatetime==2.6
pillow==10.2.0
premailer==3.10.0
prompt-toolkit==3.0.46
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
s3transfer==0.11.5
SecretStorage==3.3.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
yagmail==0.15.293
zipp==3.20.2
zope.interface==7.2
